code_evaluations_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
